fix: close audit-found correctness, contract, and pipeline gaps
Multi-phase production audit + remediation across all surfaces. All
findings were implemented as real, tested code (no stubs/mocks).

Pipeline / backend:
- PIPE-CLOCK: stamp feature-fusion physio/kinematics staleness with
  time.monotonic() — it was a time.time() epoch compared against the
  monotonic fuse() clock, so the staleness penalties were dead code.
- PIPE-REPORT: wire the SessionReport producers (interventions_triggered/
  accepted at the deliver/engage paths; record_activity/record_distraction
  from the activity-sync path). They had zero call sites, so chronotype
  trends always fell back to the HYPER proxy and task_patterns were empty.
- CONSENT-RACE: load consent state under the lock and flip _loaded only
  after the await (fixes a lost-write / stale-level race).
- REDIS-DEGRADE: degrade to the in-memory fallback on mid-session Redis
  failure (not just at connect time) and flip the degraded flag.
- ATOMIC-WRITE: unique mkstemp temp file so concurrent writers can no
  longer interleave and corrupt the destination.
- COST-BUDGET-ZERO: daily_cost_budget_usd=0 ("unlimited") now keeps
  tracking spend and never fires KILL, instead of raising and silently
  disabling the whole cost tracker + kill-switch.

Contracts (Pydantic schema is the source of truth; d.ts regenerated):
- WS COST_RESPONSE populates prompt_tokens/completion_tokens/model via the
  same shared probes GET /api/cost uses, so the two surfaces can't drift.
- no-handler TRENDS frame sends chronotype={} (the field is non-null).
- controller schedule-failure paths emit the schema-valid error='internal'.
- connected_clients / WhyDetail.error / duration_minutes docstrings aligned
  to their producers.

Frontend / desktop:
- INTERVENTION_FAILED now has consumers on every surface (desktop toast,
  WS-mode bridge, browser popup error banner) — it was broadcast with none,
  so total-mutation failures were invisible to the user.
- INTERVENTION_PROMPT cross-surface sync to the browser popup.
- start_timer overlay action drives a real countdown (was a no-op button).
- screen-share detection (CGDisplayIsCaptured / CGDisplayIsInMirrorSet)
  suppresses the always-on-top overlay while sharing; the overlay now
  targets the screen under the cursor on multi-monitor setups.

Infra / tests:
- LEETCODE-CAPS: prune adapter capabilities to the 5 the matrix can emit.
- remove the dead schema_version helper module + its test.
- Makefile typecheck uses --strict and test mirrors CI exactly; the
  ci-release-parity guard now also compares the ruff lint step.
- remove the orphaned top-level tests/ (stale, failing duplicates of
  cortex/tests/unit) and make the longitudinal aggregator test
  timezone-hermetic.

Gates: 2340 pytest pass / mypy --strict clean / ruff clean / extension +
vscode tsc+vitest green / schema+token drift + ci-release parity green.

+++ cortex/libs/schemas/ws_message_types.py
@@ -328,7 +328,7 @@ class MessageType(str, Enum):  # noqa: UP042 — pydantic-to-typescript requires
     keyboard shortcut, or the existing F26 frustration-spiral path).
     Payload mirrors :class:`cortex.libs.schemas.realtime.QuietModeState`:
     ``{kind: "snooze_15" | "quiet_session" | "pause" | "off",
-    duration_minutes: float | None, ends_at: float | None,
+    duration_minutes: int | None, ends_at: float | None,
     source: "dashboard" | "overlay" | "tray" | "shortcut" | "popup" |
     "vscode" | "os_notification" | "settings_sync" | "daemon" |
     "daemon_decay"}`` where ``ends_at`` is a unix timestamp (seconds)."""
